+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 
 ### 2020-03-06
 - Added script to generate an access token for Google APIs
+
+### 2020-04-18
+- KafkaQuery - Ability to get kafka messages between a specific start time and end time
+- KafkaPublisher - Publishes the content of input file as a single message or publishes each line from the file, which
+should be valid JSON, as a message